+++ src/main/java/com/sk89q/worldedit/forge/compat/rotation/RotationMapping.java
@@ -1,4 +1,6 @@
-package com.sk89q.worldedit.forge.compat;
+package com.sk89q.worldedit.forge.compat.rotation;
+
+import com.sk89q.worldedit.forge.compat.rotation.types.RotationBase;
 
 /** Holds the rotation type and mapping data for a block. */
 public class RotationMapping {
