+++ src/main/java/com/sk89q/worldedit/forge/compat/rotation/RotationType.java
@@ -1,4 +1,4 @@
-package com.sk89q.worldedit.forge.compat;
+package com.sk89q.worldedit.forge.compat.rotation;
 
 /**
  * Types of rotation mappings for modded blocks.
