+++ src/main/java/com/sk89q/worldedit/forge/compat/rotation/types/RotationBase.java
@@ -1,4 +1,4 @@
-package com.sk89q.worldedit.forge.compat;
+package com.sk89q.worldedit.forge.compat.rotation.types;
 
 import com.sk89q.worldedit.math.transform.AffineTransform;
