+++ src/main/java/com/sk89q/worldedit/forge/compat/rotation/types/TrapdoorRotation.java
@@ -1,5 +1,6 @@
-package com.sk89q.worldedit.forge.compat;
+package com.sk89q.worldedit.forge.compat.rotation.types;
 
+import com.sk89q.worldedit.forge.compat.rotation.RotationUtils;
 import com.sk89q.worldedit.math.transform.AffineTransform;
 
 /** Rotation handling for trapdoors supporting open/top states. */
